Safari-611.1.10/LayoutTests/fast/css/DOMQuad-points-SameObject-expected.txt

Tests that DOMQuad.p1 / p2 / p3 / p4 is [SameObject].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS quad.p1.foo is 1
PASS quad.p2.foo is 2
PASS quad.p3.foo is 3
PASS quad.p4.foo is 4
PASS quad.p1.foo is 1
PASS quad.p2.foo is 2
PASS quad.p3.foo is 3
PASS quad.p4.foo is 4
PASS quad.p1.foo is 1
PASS quad.p2.foo is 2
PASS quad.p3.foo is 3
PASS quad.p4.foo is 4
PASS successfullyParsed is true

TEST COMPLETE
